refactored server directory
